Do not return partitions which do not need election when all partitions is requested
